Safari-611.1.10/LayoutTests/http/tests/canvas/philip/tests/security.drawImage.image.html

<!DOCTYPE html>
<title>Canvas test: security.drawImage.image</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="toDataURL.jpeg.quality.outsiderange.html" title="toDataURL.jpeg.quality.outsiderange">
<link rel="next" href="security.drawImage.canvas.html" title="security.drawImage.canvas">
<body class="show_output">
<p>
 <a href="toDataURL.jpeg.quality.outsiderange.html" accesskey="p" title="[p] toDataURL.jpeg.quality.outsiderange">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="security.drawImage.canvas.html" accesskey="n" title="[n] security.drawImage.canvas">&gt;</a>
<h1><a href="index.security.html">security</a>.<a href="index.security.drawImage.html">drawImage</a>.image</h1>
<p class="desc">drawImage of different-origin image makes the canvas origin-unclean</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.security.drawImage.image">security.drawImage.image</a>
<li><a href="spec.html#testrefs.security.toDataURL">security.toDataURL</a>
<li><a href="spec.html#testrefs.security.getImageData">security.getImageData</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.drawImage(document.getElementById('yellow.png'), 0, 0);
try { var _thrown = false; canvas.toDataURL(); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: canvas.toDataURL()"); }
try { var _thrown = false; ctx.getImageData(0, 0, 1, 1); } catch (e) { _thrown = true; } finally { _assert(_thrown, "should throw exception: ctx.getImageData(0, 0, 1, 1)"); }


});
</script>
<img src="http://localhost:8000/canvas/philip/images/yellow.png" id="yellow.png" class="resource">
