Safari-611.1.10/LayoutTests/http/tests/canvas/philip/tests/security.drawImage.image-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
< [index] >
security.drawImage.image
drawImage of different-origin image makes the canvas origin-unclean
References: security.drawImage.image security.toDataURL security.getImageData
Actual output:
Passed
